[feature]-[53]-[semver:minor]-Survey yüzdesi düzeltildi.

+++ app/src/main/java/com/android/surveysaurus/model/DataPercent.kt
@@ -4,5 +4,5 @@ data class DataPercent (
     val choices: ArrayList<String>,
     val counts: ArrayList<Int>,
     val question: String,
-    val percent: ArrayList<Int>
+    val percent: ArrayList<Float>
         )
